jekyll: remove more of the old linter architecture
The front matter validation is being handled by Pydantic when I read the
source file.

For #1196
