+++ Gemfile
@@ -4,7 +4,6 @@ gem 'bit-struct', '~> 0.17'
 gem 'html-proofer', '~> 5'
 gem 'jekyll', '~> 4'
 gem 'jekyll-include-cache', '~> 0.2'
-gem 'json-schema', '~> 6'
 gem 'logger', '~> 1.7'
 gem 'minify_html', '~> 0.18'
 gem 'nokogiri', '~> 1.19'
