--- front-matter.json
@@ -1,114 +1,0 @@
-{
-    "type": "object",
-    "required": ["title", "layout"],
-    "properties": {
-        "layout": {
-            "description": "Which layout template to use.",
-            "type": "string",
-            "enum": ["page", "post", "til"]
-        },
-        "title": {
-            "description": "The title of the page or the post.",
-            "type": "string"
-        },
-        "date": {
-            "description": "When this page or post was created.",
-            "type": "string",
-            "pattern": "\\d\\d\\d\\d-\\d\\d-\\d\\d( \\d\\d:\\d\\d:\\d\\d [+\\-]\\d\\d\\d\\d)?",
-            "format": "date"
-        },
-        "date_updated": {
-            "description": "When this page or post was last updated.",
-            "type": "string",
-            "pattern": "\\d\\d\\d\\d-\\d\\d-\\d\\d( \\d\\d:\\d\\d:\\d\\d [+\\-]\\d\\d\\d\\d)?",
-            "format": "date"
-        },
-        "summary": {
-            "description": "A short description of the page or post. This is used for the <meta> tags and social media previews.",
-            "type": "string"
-        },
-        "tags": {
-            "description": "A space-separated list of tags for this page or post.",
-            "type": "array",
-            "items": {"type": "string"}
-        },
-        "link": {
-            "description": "An external page that this page should link to, like a link blog. This is used on the HTML version of the page, and where clicking on the RSS feed entry will take the reader.",
-            "type": "string"
-        },
-        "link_direct": {
-            "description": "Whether the articles index should link directly to the external page.",
-            "type": "boolean"
-        },
-        "canonical_url": {
-            "description": "An external URL that is the canonical copy of this page, used for external posts I've copied to this site.",
-            "type": "string"
-        },
-        "colors": {
-            "description": "Colours used to theme/style this individual page or post.",
-            "type": "object",
-            "properties": {
-                "css_light": {
-                    "pattern": "#[0-9a-fA-F]{6}",
-                    "type": "string"
-                },
-                "css_dark": {
-                    "pattern": "#[0-9a-fA-F]{6}",
-                    "type": "string"
-                },
-                "index_light": {
-                    "pattern": "#[0-9a-fA-F]{6}",
-                    "type": "string"
-                },
-                "index_dark": {
-                    "pattern": "#[0-9a-fA-F]{6}",
-                    "type": "string"
-                }
-            }
-        },
-        "index": {
-            "description": "Affects how an individual page or post appears in the site-wide indexes.",
-            "type": "object",
-            "properties": {
-                "exclude": {
-                    "description": "Whether to exclude this post from the site-wide index.",
-                    "type": "boolean"
-                },
-                "feature": {
-                    "description": "Whether to feature this article on the homepage",
-                    "type": "boolean"
-                }
-            },
-            "additionalProperties": false
-        },
-        "part_of": {
-            "description": "An anchor tag that describes what section of the site this is part of",
-            "type": "object",
-            "properties": {
-                "url": {
-                    "description": "The URL of the site section",
-                    "type": "string"
-                },
-                "label": {
-                    "description": "The description of this section",
-                    "type": "string"
-                }
-            },
-            "additionalProperties": false
-        },
-        "nav_section": {
-            "description": "Which section of the site this page/post belongs to.",
-            "type": "string",
-            "enum": ["articles", "til", "contact", "subscribe", "tags"]
-        },
-        "card_attribution": {
-            "description": "Any attribution information about the photo used for the card.",
-            "type": "string"
-        },
-        "excerpt_separator": {
-            "description": "Unused except to suppress a Jekyll warning",
-            "type": "string"
-        }
-    },
-    "additionalProperties": false
-}
